fix: address round 5 PR review suggestions
