+++ src/components/pwa/IosInstallModal.tsx
@@ -65,7 +65,7 @@ export function IosInstallModal({ onClose }: IosInstallModalProps) {
                     <button
                         onClick={onClose}
                         aria-label="닫기"
-                        className="text-secondary-500 hover:text-secondary-300 focus-visible:ring-primary-500 text-xl leading-none transition-colors focus-visible:ring-2 focus-visible:outline-none"
+                        className="text-secondary-500 hover:text-secondary-300 focus-visible:ring-primary-500 text-xl leading-none transition-colors focus-visible:ring-1 focus-visible:outline-none"
                     >
                         ×
                     </button>
