+++ src/components/symbol-page/hooks/useAnalysis.ts
@@ -24,7 +24,6 @@ import {
     tryAcquireReanalyzeCooldown,
 } from '@/infrastructure/market/reanalyzeCooldown';
 import { sleep } from '@/components/symbol-page/utils/sleep';
-import { PWA_TRIGGER_EVENT } from '@/lib/pwaEvents';
 
 interface AnalyzeMutationVariables {
     symbol: string;
@@ -73,6 +72,8 @@ export interface CooldownNotice {
 
 interface UseAnalysisResult {
     analysis: AnalysisResponse;
+    /** 새 분석이 완료됐을 때만 값이 세팅됨. initialAnalysis 기반 초기 로드엔 null. */
+    analysisResult: AnalysisResponse | null;
     isAnalyzing: boolean;
     analysisError: string | null;
     handleReanalyze: () => void;
@@ -149,7 +150,6 @@ export function useAnalysis({
                 if (data.status === 'cached') {
                     currentJobIdRef.current = null;
                     setAnalysisResult(data.result);
-                    window.dispatchEvent(new CustomEvent(PWA_TRIGGER_EVENT));
                     // 캐시 히트 = 분석 완료 → force 경로만 쿨다운 시작
                     if (variables.force) {
                         setReanalyzeCooldownMs(REANALYZE_COOLDOWN_MS);
@@ -253,9 +253,6 @@ export function useAnalysis({
                     if (result.status === 'done') {
                         currentJobIdRef.current = null;
                         setAnalysisResult(result.result);
-                        window.dispatchEvent(
-                            new CustomEvent(PWA_TRIGGER_EVENT)
-                        );
                         if (lastForceRef.current) {
                             setReanalyzeCooldownMs(REANALYZE_COOLDOWN_MS);
                         }
@@ -384,6 +381,7 @@ export function useAnalysis({
 
     return {
         analysis,
+        analysisResult,
         isAnalyzing,
         analysisError,
         handleReanalyze,
